+++ features/main/calendar/atoms/index.ts
@@ -1,3 +1,3 @@
-export { CalendarHeader } from './calendar-header';
 export { CalendarItem } from './calendar-item';
 export { DayLabels } from './day-labels';
+export { MoveMonthButton } from './move-month-button';
